# shared/api/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  piemdm_internal_model.Table:
    properties:
      DisplayMode:
        description: 展示模式：List 列表, Tree 树形
        enum:
        - List
        - Tree
        type: string
      ParentField:
        description: 父表关联字段
        maxLength: 64
        type: string
      ParentTable:
        description: 父表Code（仅当TableType=Item时有效，不带t_前缀）
        maxLength: 64
        type: string
      SelfField:
        description: 本表关联字段
        maxLength: 64
        type: string
      TableType:
        description: 表类型：Entity 主表实体, Item 行项目
        enum:
        - Entity
        - Item
        type: string
      code:
        description: 表名,英文
        maxLength: 64
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        maxLength: 255
        type: string
      id:
        type: integer
      name:
        description: 名称，可以中文
        maxLength: 128
        type: string
      sort:
        maximum: 9999
        type: integer
      status:
        description: 状态：Normal 正常 Frozen 已冻结 Deleted 已删除
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
    required:
    - code
    - name
    type: object
...
  piemdm_internal_model.Webhook:
    properties:
      contentType:
        description: |-
          请求方式：POST，GET
          数据类型：plain，application/json，application/xxx-form-urlencoded，application/xml
          Data：string
          Headers；[]Key，Value
          Response示例
          Basic Auth 用户名
          Basic Auth 密码
        maxLength: 128
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        maxLength: 255
        type: string
      events:
        description: 监控什么事件
        maxLength: 256
        type: string
      id:
        type: integer
      secret:
        description: AppSecret
        maxLength: 128
        type: string
      status:
        description: 状态：Normal 正常 Frozen 已冻结 Deleted 已删除
        type: string
      tableCode:
        description: 主数据Code
        maxLength: 64
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
      url:
        description: 调用Url
        maxLength: 256
        type: string
      username:
        description: 用户名
        maxLength: 64
        type: string
    required:
    - events
    - secret
    - url
    type: object
...
host: localhost:8787
info:
  contact:
    email: support@piemdm.com
    name: API Support
  description: PieMDM Master Data Management System API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PieMDM API
  version: "1.0"
paths:
...
  /admin/tables:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 15
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 表代码
        in: query
        name: code
        type: string
      - description: 开始日期
        in: query
        name: startDate
        type: string
      - description: 结束日期
        in: query
        name: endDate
        type: string
      - description: 关联关系
        in: query
        name: relation
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/piemdm_internal_model.Table'
            type: array
      summary: 获取表列表
      tags:
      - 表管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 表信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/piemdm_internal_model.Table'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.Table'
      summary: 创建表
      tags:
      - 表管理
  /admin/tables/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 表ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.Table'
      summary: 获取表详情
      tags:
      - 表管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 表ID
        in: path
        name: id
        required: true
        type: integer
      - description: 表信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/piemdm_internal_model.Table'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 更新表
      tags:
      - 表管理
  /admin/tables/batch:
    put:
      consumes:
      - application/json
      parameters:
      - description: 批量更新请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 批量更新表状态
      tags:
      - 表管理
  /admin/tables/batch_delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 批量删除请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 批量删除表
      tags:
      - 表管理
...
  /admin/webhooks:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 15
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 表代码
        in: query
        name: tableCode
        type: string
      - description: 开始日期
        in: query
        name: startDate
        type: string
      - description: 结束日期
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/piemdm_internal_model.Webhook'
            type: array
      summary: 获取Webhook列表
      tags:
      - Webhook管理
    post:
      consumes:
      - application/json
      parameters:
      - description: Webhook信息
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.Webhook'
      summary: 创建Webhook
      tags:
      - Webhook管理
  /admin/webhooks/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: WebhookID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 删除Webhook
      tags:
      - Webhook管理
    get:
      consumes:
      - application/json
      parameters:
      - description: WebhookID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.Webhook'
      summary: 获取Webhook详情
      tags:
      - Webhook管理
    put:
      consumes:
      - application/json
      parameters:
      - description: WebhookID
        in: path
        name: id
        required: true
        type: integer
      - description: Webhook信息
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 更新Webhook
      tags:
      - Webhook管理
  /admin/webhooks/batch:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 批量删除Webhook
      tags:
      - Webhook管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 批量更新请求
        in: body
        name: data
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 批量更新Webhook状态
      tags:
      - Webhook管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"